table migration

+++ week6/by-its-cover/models/favorite.js
@@ -3,7 +3,7 @@ module.exports = function(sequelize, DataTypes) {
   var favorite = sequelize.define('favorite', {
     title: DataTypes.STRING,
     author: DataTypes.STRING,
-    isbn: DataTypes.INTEGER,
+    isbn: DataTypes.BIGINT,
     rating: DataTypes.INTEGER,
     pageCount: DataTypes.INTEGER,
     description: DataTypes.STRING,
